% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Ca_Mu}
\alias{Ca_Mu}
\title{The expected mu vector}
\usage{
Ca_Mu(y, nn, X, family, beta)
}
\arguments{
\item{y}{response variable vector}

\item{nn}{weight vector}

\item{X}{covariate matrix}

\item{family}{the distribution family}

\item{beta}{parameter estimate vector}
}
\value{
expected mu vector
}
\description{
The expected mu vector
}
\examples{
 \dontrun{ramu <- Ca_Mu(phenotype,weight,  cov, family, ra)}

}
