% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Ca_Wii}
\alias{Ca_Wii}
\title{The diagonal elements of the W matrix}
\usage{
Ca_Wii(y, nn, family, mu, beta, a)
}
\arguments{
\item{y}{response variable vector}

\item{nn}{weight vector}

\item{family}{the distribution family}

\item{mu}{expected vector}

\item{beta}{estimated parameter values}

\item{a}{theta parameter}
}
\value{
Vector of diagonal elements of the W matrix (Wii)
}
\description{
The diagonal elements of the W matrix
}
\examples{
 \dontrun{rawii <- Ca_Wii(phenotype,weight,  family, ramu,ra,theta)}

}
