% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_test.R
\name{define_Ga}
\alias{define_Ga}
\title{Conversion of genoSnpMatrix to a genotype matrix}
\usage{
define_Ga(genoSnpMatrix, standardization = TRUE)
}
\arguments{
\item{genoSnpMatrix}{a genotype matrix in SnpMatrix format}

\item{standardization}{logical, indicating whether to standardize the matrix}
}
\value{
a genotype matrix with 0, 1, 2 encoding and optionally standardized
}
\description{
This function takes a genoSnpMatrix in SnpMatrix format and converts it to a
genotype matrix with 0, 1, 2 encoding. By default, it standardizes the matrix
using the formula (x - 2*p) / sqrt(2*p*(1-p)), where x is the encoded numeric
value in each column and p is the minor allele frequency (maf) in each column.
}
\examples{
\dontrun{
load("./inst/PhenoGenoMap.RData")
G <- snp_Data$genotypes
Ga <- define_Ga(G)}

}
