% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_test.R
\name{epi.snp.estimate}
\alias{epi.snp.estimate}
\title{Filtering and parameter estimation for significant SNPs in 2D Scaning}
\usage{
epi.snp.estimate(
  significantASNP = NULL,
  phenotype = phenotype,
  covariates = NULL,
  environment = NULL,
  weights = NULL,
  data = NULL,
  theta = NULL,
  genodata = genodata,
  map = map,
  Pvalues = Pvalues,
  Bonferr = 0.5,
  family = "gaussian",
  lasso = FALSE,
  selection = FALSE,
  estimate = FALSE,
  GEI_random = TRUE
)
}
\arguments{
\item{significantASNP}{a vector of significant SNP names from 1D scanning}

\item{phenotype}{a vector of phenotype values}

\item{covariates}{a matrix of covariates (PCA)}

\item{environment}{a vector of environmental factors}

\item{weights}{a vector of weights}

\item{data}{a data frame containing the input data}

\item{theta}{the theta parameter}

\item{genodata}{a genotype matrix}

\item{map}{the map file}

\item{Pvalues}{a vector of 2D p-values}

\item{Bonferr}{bonferroni correction factor, 0.5/(M * (M - 1)/2)}

\item{family}{the distribution family}

\item{lasso}{logical, LASSO method}

\item{selection}{logical, forward selection method}

\item{estimate}{logical, estimate parameters for significant markers}

\item{GEI_random}{logical, gene-environment interaction effects as random effects}
}
\value{
A list containing information about significant SNP markers from 1D and 2D scans,
        covariates coefficients, genotype coefficients, GEI coefficients, GEI VarCorr, and AIC.
}
\description{
This function filters the results from an A-A epistatic 2D scaning to obtain
significant SNP markers and allows for parameter estimation.
}
\examples{
 \dontrun{
sig <- c("SNP25","SNP67")
raae1 <- epi.snp.estimate(significantASNP = sig,phenotype = y, covariates = cov,
                          environment = env, genodata = Ga,map = map,Pvalues = p2,
                          Bonferr = 0.5,family = "binomial",selection = T,estimate = T)}

}
