% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_test.R
\name{epi.snp.score}
\alias{epi.snp.score}
\title{2D scanning for additive-additive epistasis and epistasis-by-environment SNPs}
\usage{
epi.snp.score(
  significantASNP = NULL,
  phenotype = phenotype,
  covariates = NULL,
  environment = NULL,
  weights = NULL,
  data = NULL,
  threshold = 0.01,
  theta = NULL,
  genodata = genodata,
  map = map,
  family = "gaussian",
  parallel = FALSE,
  num_threads = 2,
  filename = NULL
)
}
\arguments{
\item{significantASNP}{a vector of significant SNP names from a 1D scanning}

\item{phenotype}{a vector of phenotype values}

\item{covariates}{a matrix of covariates (PCA)}

\item{environment}{a vector of environmental factors}

\item{weights}{a vector of weights}

\item{data}{a data frame containing the input data}

\item{threshold}{the threshold for precise calculation and estimation}

\item{theta}{the theta parameter when family is "negative.binomial"}

\item{genodata}{a genotype matrix}

\item{map}{the map file}

\item{family}{the distribution family}

\item{parallel}{logical, parallel computation}

\item{num_threads}{the number of threads for parallel computation}

\item{filename}{the filename for output}
}
\value{
A list containing the 1D scan results for additive-additive epistasis,
       including cnromosome, SNP name, position,p-values, and score values.
}
\description{
This function conducts a two-dimensional scan for on a SNP matrix based on the
given parameters and significant SNPs from a 1D scanning.
}
\examples{
\dontrun{
sig <- c("SNP25","SNP67")
raae <- epi.snp.score(significantASNP = sig, phenotype = y, covariates = cov,
                     environment = env,genodata = Ga,map = map,family = "binomial",
                     parallel = T, num_threads = 4,filename = "gwasp2.txt")}

}
