% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{forward_selection}
\alias{forward_selection}
\title{Forward selection method}
\usage{
forward_selection(
  y,
  nn,
  X,
  G,
  family,
  envir,
  threshold,
  a,
  x_indices,
  y_indices
)
}
\arguments{
\item{y}{response variable vector}

\item{nn}{weight vector}

\item{X}{covariate matrix}

\item{G}{genotype matrix}

\item{family}{the distribution family}

\item{envir}{number of environmental factors}

\item{threshold}{selection threshold}

\item{a}{theta parameter}

\item{x_indices}{indices of features to be selected}

\item{y_indices}{indices of features to be retained}
}
\value{
Vector of indices of selected significant features (x_selected)
}
\description{
Forward selection method
}
\examples{
 \dontrun{fin <- forward_selection(phenotype1,weight, cov, Gaa,family,
          envir,thr, theta, fin_s, fin_l)}

}
