% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateSubmatrix}
\alias{generateSubmatrix}
\title{Generate a submatrix by replicating the matrix based on envir}
\usage{
generateSubmatrix(X, envir)
}
\arguments{
\item{X}{input matrix}

\item{envir}{number of environmental factors}
}
\value{
Transformed matrix with replicated blocks (Xs)
}
\description{
Generate a submatrix by replicating the matrix based on envir
}
\examples{
 \dontrun{generateSubmatrix(gf, envir)}

}
