% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{glm_NR}
\alias{glm_NR}
\title{Parameters estimation using the Newton-Raphson method}
\usage{
glm_NR(y, nn, X, family, a)
}
\arguments{
\item{y}{response variable vector}

\item{nn}{weight vector}

\item{X}{covariate matrix}

\item{family}{the distribution family}

\item{a}{theta parameter}
}
\value{
estimated parameter values (beta)
}
\description{
Parameters estimation using the Newton-Raphson method
}
\examples{
 \dontrun{ra <- glm_NR(phenotype,weight, cov, family,theta)}

}
