% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_test.R
\name{logit}
\alias{logit}
\title{The logit transformation}
\usage{
logit(x)
}
\arguments{
\item{x}{A numeric value for the expected value mu}
}
\value{
A numeric value for the linear predictor yita
}
\description{
This function takes the expected value mu (in the range of 0 to 1) and transforms
it to the linear predictor yita in the real number domain.
}
\examples{
\dontrun{logit(0.5)}

}
