% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{score_test}
\alias{score_test}
\title{Score test for model testing}
\usage{
score_test(
  y,
  mu,
  Wii,
  G,
  proj,
  beta,
  family,
  envir,
  b,
  a,
  epistasis_test = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{y}{response variable vector}

\item{mu}{expected vector}

\item{Wii}{vector of diagonal elements of the W matrix}

\item{G}{genotype matrix}

\item{proj}{projection matrix}

\item{beta}{estimated parameter values}

\item{family}{the distribution family}

\item{envir}{number of environmental factors}

\item{b}{threshold for computation}

\item{a}{theta parameter}

\item{epistasis_test}{flag for conducting 2D scan}

\item{parallel}{flag for parallel computation}
}
\value{
List containing 'x2' (score value) and 'p_value' (p-value)
}
\description{
Score test for model testing
}
\examples{
 \dontrun{ra1 <- score_test(phenotype, ramu, rawii, genodata,proj, ra,family,
         envir,threshold,theta, epistasis_test = TRUE, parallel = parallel)}

}
