% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_test.R
\name{single.snp.estimate}
\alias{single.snp.estimate}
\title{Filtering and parameter estimation for significant SNPs in 1D Scaning}
\usage{
single.snp.estimate(
  phenotype = phenotype,
  covariates = NULL,
  environment = NULL,
  weights = NULL,
  data = NULL,
  theta = NULL,
  genodata = genodata,
  map = map,
  Pvalues = Pvalues,
  Bonferr = 0.5,
  family = "gaussian",
  lasso = FALSE,
  selection = FALSE,
  estimate = FALSE,
  GEI_random = TRUE
)
}
\arguments{
\item{phenotype}{a vector of phenotype values}

\item{covariates}{a matrix of covariates}

\item{environment}{a vector of environmental}

\item{weights}{a vector of weights}

\item{data}{a data frame containing the input data}

\item{theta}{the theta parameter}

\item{genodata}{a genotype matrix}

\item{map}{the map file}

\item{Pvalues}{a vector of 1D p-values}

\item{Bonferr}{bonferroni correction factor, 0.5/M (M is the number of SNPs)}

\item{family}{the distribution family}

\item{lasso}{logical, LASSO method}

\item{selection}{logical, forward selection method}

\item{estimate}{logical, estimating parameters for significant markers}

\item{GEI_random}{logical, gene-environment interaction effects as random effects}
}
\value{
A list containing information about 1D significant SNPs, covariates coefficients,
        genotype coefficients, GEI coefficients, GEI VarCorr, and model AIC.
}
\description{
This function filters the results from an additive 1D scaning to obtain significant
SNP markers and allows for parameter estimation.
}
\examples{
 \dontrun{rae1 <- single.snp.estimate(phenotype = y, covariates = cov, environment = env,
                            genodata = Ga,map = map,Pvalues = p1, Bonferr = 0.5,
                            family = "binomial",selection = T,estimate = T)}

}
