% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_test.R
\name{single.snp.score}
\alias{single.snp.score}
\title{1D scanning for additive and additive-by-environment SNPs}
\usage{
single.snp.score(
  phenotype = phenotype,
  covariates = NULL,
  environment = NULL,
  weights = NULL,
  data = NULL,
  threshold = 0.01,
  theta = NULL,
  genodata = genodata,
  map = map,
  family = "gaussian",
  parallel = FALSE,
  num_threads = 2,
  filename = NULL
)
}
\arguments{
\item{phenotype}{a vector of phenotype values}

\item{covariates}{a matrix of covariates (PCA)}

\item{environment}{a vector of environmental factors}

\item{weights}{a vector of weights}

\item{data}{a data frame containing the input data}

\item{threshold}{the threshold for precise calculation and estimation}

\item{theta}{the theta parameter when family is "negative.binomial"}

\item{genodata}{a genotype matrix}

\item{map}{the map file}

\item{family}{the distribution family}

\item{parallel}{logical, parallel computation}

\item{num_threads}{the number of threads for parallel computation}

\item{filename}{the filename for output}
}
\value{
A list containing the 1D scan results, including chromosome, SNP name,
        position, score value and p-value.
}
\description{
This function conducts a one-dimensional scan on SNPs based on the given parameters.
}
\examples{
\dontrun{rae <- single.snp.score(phenotype = y,covariates = cov, environment = env,
                         genodata = Ga,map = map,family = "binomial")}

}
